Creating a lib module of plotter funcs
